ass_rasterizer: fix ilog2 impl on MSVC
Fix type of res variable.

Note that result is undefined if n==0, same as __builtin_clz. I assume
this contract is respected in all callsites.

+++ libass/ass_rasterizer.c
@@ -36,7 +36,7 @@ static inline int ilog2(uint32_t n)
 #ifdef __GNUC__
     return __builtin_clz(n) ^ 31;
 #elif defined(_MSC_VER)
-    int res;
+    unsigned long res;
     _BitScanReverse(&res, n);
     return res;
 #else
